#12 Add warning for compiling examples
"Older" Nim versions such as 2.0.8 do not deal with gcc's
and clang's typed pointers.
That is a feature that these compilers only incorporated
last year or so, which causes a lot of nim generated C code
to on longer compile.

Therefore a warning is required to inform users of that.